consolidated mock
