+++ test/integration/mapper.py
@@ -1,0 +1,7 @@
+from test import mock
+from autofit import mapper
+
+
+class TestCase(object):
+    def test(self):
+        mapper.PriorModel(mock)
