+++ test/test_equality.py
@@ -5,7 +5,8 @@ import pytest
 import autofit.mapper.model
 import autofit.optimize.non_linear.multi_nest
 import autofit.optimize.non_linear.non_linear
-from autofit import mock
+import test.mock
+from test import mock
 from autofit.mapper import model_mapper as mm
 from autofit.mapper import prior as p
 from autofit.mapper import prior_model as pm
@@ -14,7 +15,7 @@ from autofit.optimize import non_linear
 
 @pytest.fixture(name="prior_model")
 def make_prior_model():
-    return pm.PriorModel(mock.GeometryProfile)
+    return pm.PriorModel(test.mock.GeometryProfile)
 
 
 class TestCase(object):
@@ -48,7 +49,7 @@ class TestCase(object):
 
     def test_non_trivial_equality(self):
         model_mapper = mm.ModelMapper()
-        model_mapper.galaxy = mock.GalaxyModel(light_profile=mock.GeometryProfile, mass_profile=mock.GeometryProfile)
+        model_mapper.galaxy = test.mock.GalaxyModel(light_profile=test.mock.GeometryProfile, mass_profile=test.mock.GeometryProfile)
         model_mapper_copy = deepcopy(model_mapper)
 
         assert model_mapper == model_mapper_copy
@@ -59,7 +60,7 @@ class TestCase(object):
 
     def test_model_instance_equality(self):
         model_instance = autofit.mapper.model.ModelInstance()
-        model_instance.profile = mock.GeometryProfile()
+        model_instance.profile = test.mock.GeometryProfile()
         model_instance_copy = deepcopy(model_instance)
 
         assert model_instance == model_instance_copy
@@ -70,7 +71,7 @@ class TestCase(object):
 
     def test_non_linear_equality(self):
         nlo = autofit.optimize.non_linear.non_linear.NonLinearOptimizer("phase name")
-        nlo.variable.profile = mock.GeometryProfile
+        nlo.variable.profile = test.mock.GeometryProfile
         nlo_copy = deepcopy(nlo)
 
         assert nlo_copy == nlo
