+++ test/tools/test_phase_property.py
@@ -4,8 +4,9 @@ import pytest
 
 import autofit.mapper.prior_model
 import autofit.optimize.non_linear.non_linear
+import test.mock
 from autofit import conf
-from autofit import mock
+from test import mock
 from autofit.mapper import model_mapper as mm
 from autofit.optimize import non_linear
 from autofit.tools import phase as ph
@@ -68,7 +69,7 @@ class TestPhasePropertyList(object):
     #     assert list_phase.prop == objects
 
     def test_classes(self, list_phase):
-        objects = [mock.GalaxyModel(), mock.GalaxyModel()]
+        objects = [test.mock.GalaxyModel(), test.mock.GalaxyModel()]
 
         list_phase.prop = objects
 
@@ -84,7 +85,7 @@ class TestPhasePropertyList(object):
         assert list_phase.prop == objects
 
     def test_mix(self, list_phase):
-        objects = [mock.GalaxyModel(), mock.Galaxy()]
+        objects = [test.mock.GalaxyModel(), test.mock.Galaxy()]
 
         list_phase.prop = objects
 
@@ -92,17 +93,17 @@ class TestPhasePropertyList(object):
         assert list_phase.prop == objects
 
     def test_set_item(self, list_phase):
-        galaxy_prior_0 = mock.GalaxyModel()
-        objects = [galaxy_prior_0, mock.Galaxy()]
+        galaxy_prior_0 = test.mock.GalaxyModel()
+        objects = [galaxy_prior_0, test.mock.Galaxy()]
 
         list_phase.prop = objects
 
-        galaxy_prior_1 = mock.GalaxyModel()
+        galaxy_prior_1 = test.mock.GalaxyModel()
         list_phase.prop[1] = galaxy_prior_1
 
         assert list_phase.variable.prop == [galaxy_prior_0, galaxy_prior_1]
 
-        galaxy = mock.Galaxy()
+        galaxy = test.mock.Galaxy()
 
         list_phase.prop[0] = galaxy
         assert list_phase.prop == [galaxy, galaxy_prior_1]
@@ -110,7 +111,7 @@ class TestPhasePropertyList(object):
 
 class TestPhasePropertyCollectionAttributes(object):
     def test_set_list_as_dict(self, list_phase):
-        galaxy_model = mock.GalaxyModel()
+        galaxy_model = test.mock.GalaxyModel()
         list_phase.prop = dict(one=galaxy_model)
 
         assert len(list_phase.prop) == 1
@@ -118,9 +119,9 @@ class TestPhasePropertyCollectionAttributes(object):
         assert list_phase.prop.one == galaxy_model
 
     def test_override_property(self, list_phase):
-        galaxy_model = mock.GalaxyModel()
+        galaxy_model = test.mock.GalaxyModel()
 
-        list_phase.prop = dict(one=mock.GalaxyModel())
+        list_phase.prop = dict(one=test.mock.GalaxyModel())
 
         list_phase.prop.one = galaxy_model
 
@@ -128,23 +129,23 @@ class TestPhasePropertyCollectionAttributes(object):
         assert list_phase.prop.one == galaxy_model
 
     def test_named_list_items(self, list_phase):
-        galaxy_model = mock.GalaxyModel()
+        galaxy_model = test.mock.GalaxyModel()
         list_phase.prop = [galaxy_model]
 
         # noinspection PyUnresolvedReferences
         assert getattr(list_phase.prop, "0") == galaxy_model
 
     def test_mix(self, list_phase):
-        objects = dict(one=mock.GalaxyModel(), two=mock.Galaxy())
+        objects = dict(one=test.mock.GalaxyModel(), two=test.mock.Galaxy())
 
         list_phase.prop = objects
 
-        list_phase.prop.one = mock.Galaxy()
+        list_phase.prop.one = test.mock.Galaxy()
 
         assert len(list_phase.variable.prop) == 2
 
     def test_named_attributes_in_variable(self, list_phase):
-        galaxy_model = mock.GalaxyModel(variable_redshift=True)
+        galaxy_model = test.mock.GalaxyModel(variable_redshift=True)
         list_phase.prop = dict(one=galaxy_model)
 
         assert list_phase.variable.prior_count == 1
@@ -156,11 +157,11 @@ class TestPhasePropertyCollectionAttributes(object):
         assert len(instance.prop) == 1
 
     def test_named_attributes_in_variable_override(self, list_phase):
-        list_phase.prop = dict(one=mock.GalaxyModel())
+        list_phase.prop = dict(one=test.mock.GalaxyModel())
 
         assert list_phase.variable.prior_count == 0
 
-        galaxy_model = mock.GalaxyModel(variable_redshift=True)
+        galaxy_model = test.mock.GalaxyModel(variable_redshift=True)
 
         list_phase.prop.one = galaxy_model
 
@@ -173,14 +174,14 @@ class TestPhasePropertyCollectionAttributes(object):
         assert len(instance.prop) == 1
 
     def test_named_attributes_in_constant(self, list_phase):
-        galaxy = mock.Galaxy()
+        galaxy = test.mock.Galaxy()
         list_phase.prop = dict(one=galaxy)
 
         assert list_phase.variable.prior_count == 0
         assert list_phase.variable.prop.one == galaxy
 
     def test_singular_model_info(self, list_phase):
-        galaxy_model = mock.GalaxyModel(variable_redshift=True)
+        galaxy_model = test.mock.GalaxyModel(variable_redshift=True)
         list_phase.prop = dict(one=galaxy_model)
 
         assert list_phase.variable.prop.one == galaxy_model
@@ -194,8 +195,8 @@ class TestPhasePropertyCollectionAttributes(object):
         assert len(list_phase.variable.info.split('\n')) == 7
 
     def test_shared_priors(self, list_phase):
-        list_phase.prop = dict(one=mock.GalaxyModel(variable_redshift=True),
-                               two=mock.GalaxyModel(variable_redshift=True))
+        list_phase.prop = dict(one=test.mock.GalaxyModel(variable_redshift=True),
+                               two=test.mock.GalaxyModel(variable_redshift=True))
 
         assert list_phase.variable.prior_count == 2
 
@@ -208,17 +209,17 @@ class TestPhasePropertyCollectionAttributes(object):
         list_phase.prop = dict()
 
         assert not hasattr(list_phase.prop, "one")
-        list_phase.prop = dict(one=mock.GalaxyModel(variable_redshift=True))
+        list_phase.prop = dict(one=test.mock.GalaxyModel(variable_redshift=True))
 
         assert hasattr(list_phase.prop, "one")
 
     def test_position_not_a_prior(self, list_phase):
-        list_phase.prop = [autofit.mapper.prior_model.PriorModel(mock.Galaxy)]
+        list_phase.prop = [autofit.mapper.prior_model.PriorModel(test.mock.Galaxy)]
 
         assert list_phase.variable.prior_count == 1
         assert "redshift" == list_phase.variable.prior_tuples_ordered_by_id[0][0]
 
-        prior_model = autofit.mapper.prior_model.PriorModel(mock.Galaxy)
+        prior_model = autofit.mapper.prior_model.PriorModel(test.mock.Galaxy)
         prior_model.phase_property_position = 0
 
         assert len(prior_model.constant_tuples) == 0
